+++ framework/Source/CPTTextLayer.h
@@ -10,11 +10,13 @@ extern const CGFloat kCPTTextLayerMarginWidth; ///< Margin width around the text
     NSString *text;
     CPTTextStyle *textStyle;
     NSAttributedString *attributedText;
+    CGSize maximumSize;
 }
 
 @property (readwrite, copy, nonatomic) NSString *text;
 @property (readwrite, retain, nonatomic) CPTTextStyle *textStyle;
 @property (readwrite, copy, nonatomic) NSAttributedString *attributedText;
+@property (readwrite, nonatomic) CGSize maximumSize;
 
 /// @name Initialization
 /// @{